Move tensorboard logger and bin_utils in utils
